+++ webdev/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.5.2-dev
 
+- Update SDK dependency to minimum of 2.5.0.
+- Support absolute file: URIs for libraries.
 - Depend on the latest `package:dwds`.
 
 ## 2.5.1